# swagger.yaml
openapi: 3.0.0
info:
  title: Villagers API
  version: "1.0.0"
  description: A REST API for managing Villagers with CRUD functionality
servers:
  - url: http://localhost:3000/api
paths:
  /login:
    post:
      summary: Log in as admin
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /logout:
    post:
      summary: Log out the current user
      tags:
        - Authentication
      responses:
        '200':
          description: Logout successful
  /villagers:
    get:
      summary: Retrieve all villagers
      tags:
        - Villagers
      responses:
        '200':
          description: A list of villagers
    post:
      summary: Create a new villager
      tags:
        - Villagers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                hobby:
                  type: string
                gender:
                  type: string
                age:
                  type: integer
      responses:
        '201':
          description: Villager created successfully
  /villagers/{id}:
    get:
      summary: Retrieve a single villager by ID
      tags:
        - Villagers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single villager
        '404':
          description: Villager not found
    put:
      summary: Update an existing villager
      tags:
        - Villagers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                hobby:
                  type: string
                gender:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: Villager updated successfully
        '404':
          description: Villager not found
    delete:
      summary: Delete a villager
      tags:
        - Villagers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Villager deleted successfully
        '404':
          description: Villager not found